Changed default program to simplify testing.
This program blinks the 13 pin once per second. However, it's dependent on
the current instruction set so I will only keep this while refactoring the
code because I can quickly start the simulator and see if the program still
works. Later, when I update the instruction set this will be removed.