Modified Simulator to be able to control the sketch loop from C#.
This allows me to read the serial on every tick and store each entry in a
queue without fear that I lost entries due to slow reading. The potential
problem is now writing more than the buffer size in a single tick, but that
should be easy to fix by simply increasing the buffer size.