Fixed bug when the program is written past the end of the EEPROM.
The problem was that I assumed the % operator worked for negative numbers
as in Squeak. That was incorrect, so I added a positive_modulo(..) function
that works as I expected. Since this function is needed in several places
I put it in a EEPROMUtils.h file and I took the opportunity to remove some
duplicated definitions by putting them there.