Fixed bug in UziFirmware.executeSaveProgram()
The old version was failing to write relatively large programs. The problem
was that I was inmediately writing to the EEPROM each byte read from the
serial. And since writing to the EEPROM is slow, the serial buffer was
filling up and eventually I was losing data. Now I read from the serial as
fast as possible.