Added current primitives as instruction opcodes (and removed the PRIM_CALL)
This change is very important mainly because it makes it possible for
specific primitives to use multiple bytes for arguments. Before this,
primitives would not accept arguments (all their data was supposed to come
from the stack). We still don't have any primitive using arguments, though,
but even if no primitive ever needs this functionality I think the code is
simpler now.