Fixed bug when calling PRIM_RET from the ticking script.
The problem was that the stack was unwind and the pc was set to the script
start. But since the stack frame was not initialized again, the next time
the PRIM_RET was executed, it was trying to pop an empty stack.
This error was detected because the TestScriptTickingWithExplicitReturn()
test was failing randomly. I don't know how it would behave on the actual
arduino.