Implemented support for local variables.
The syntax is similar to smalltalk in that it uses the pipe character $| to
enclose the declarations but it allows to specify the default values for
each variable.
The compiler for now simply treats them exactly like arguments so the
generated code is not really optimal.
Also, now the global variables must be declared or the compilation fails.