Fixed bug in SCRIPT_CALL instruction.
The problem was that the incorrect framePointer was being pushed on the
stack. So, when unwinding the stack I had to go one extra frame down in
order to restore the correct frame pointer. Now, I'm pushing the correct
framePointer at the call site and I removed the hack in the unwinding.