Added UziStackFrame class and UziDebugger>>#stackFrames.
I moved both UziDebugger and UziStackFrame to their own class category
'Uzi-Core-Debugging'.
Also, I fixed a bug when writing/reading floats from the serial. I was
under the impression that Float>>#asIEEE32BitWord and
Float>>#fromIEEE32Bit: were symmetrical but somehow this wasn't the case.
I don't know, maybe I messed up. Anyway, the bug was that some elements on
the stack (the return address and frame pointer) were not correct so I was
not able to correctly interpret the stack to build all the stack frames.
Now I started using FloatArray to perform the conversions and everything
works.