Changed firmware to notify errors to the client.
I replaced StackError with a new global enum Error and added a reference to
an error on the Coroutine. In the sendVMState() function I now check if the
coroutine has an error and then I reset the coroutine state.
Also, when an error is detected the coroutine's original script is stopped
to avoid an endless cascade of errors.