Removed UziSpecialVariableNode.
I simply replaced it with an UziPrimitiveCallNode. I had to change the
UziTypeChecker because when visiting primitive calls it was checking that
the number of arguments matched the selector but with special variables
they didn't match.
So now all UziPrimitives know how they affect the stack. I changed
UziPrimitive>>#primSpec and added that info. I also removed the
"isExpression" inst var because now it can be calculated by using the
stack transition.