Temporary fix for the variable declaration compilation.
I added a while back an early optimization if the variable was declared
with a literal default value. The problem is that this optimization relied
on the encoder to be able to store the default value of local variables.
The encoder does not support this (neither does the vm, actually) so the
optimization is invalid. Unfortunately I wrote a test with this exact
behavior and since I don't want to change the test for now the test will
fail until I fix this in the vm/encoder (soon, I hope).