Starting to change how we allocate memory for the VM's current program.
I changed almost all the "new" instances to use the uzi memory functions.
This is nice because now the memory usage is much more predictable.
For now this seems to work fine but it worries me that it shouldn't because
coroutines could have their stack resized and the code doesn't handle that
yet. The tests don't fail so I guess I'm missing a test for this case.