Starting to fix the problem of the stack resizing in coroutines.
As a first step, I wrote a simple implementation that just allocates a new
array each time the coroutine stack needs to grow.
It's not particularly efficient because the memory occupied by the old
array is leaked every time we resize but it works if you have enough RAM
available.