Starting to handle OUT_OF_MEMORY errors. And added READER_TIMEOUT error.
I changed the "reading" procedures to return an error code. This way if
creating a program fails for whatever reason I can notify the user.
I also changed the Coroutine to check if allocation was successful when
resizing its stack. If it fails, it automatically sets its error to
OUT_OF_MEMORY and it stops itself.
The only thing missing (I think) is to notify the user when OUT_OF_MEMORY
happens at coroutine creation.