Changed Monitor connection process.
Now it performs a simple handshake with the client. The idea is to only
allow a connection when the Monitor is disconnected and only execute
commands when it is connected. This method will allow to identify a
connection request more precisely and that will become important later when
we expose the serial port to user programs.