Changed parser to treat variable declarations as just another statement.
Before it would only allow them at the beginning of the script, now they
could go anywhere a statement is allowed.
This allowed me to refactor the UziBlockNode and remove its "locals" inst
var. I left the UziBlockNode>>#locals method, though, and it now looks for
variable declarations in its statements. This implementation would probably
fail if variables are declared in nested blocks but I'll fix that later.