Added a small change in Channel.uzi to prevent future errors.
The problem is that the library depends on some specific behavior of the VM
that might change in the future. I don't want that change to break things
here so I'm putting this here as a precaution, although I'm not sure it's
enough. At least, it tells future me that I was worried about this stuff...