Improved UziProtocol, particularly regarding timeouts and disconnections.
I added a timeout to UziSerialStream, before it could hang forever waiting
for a byte to arrive. Then I changed UziProtocol to handle timeout
exceptions correctly (depending on the case, it could mean either ignoring
the error or disconnecting).
I also added handling of the DISCONNECT_ERROR, which happens when the
firmware doesn't receive a keep alive signal in time. I added this because
if, for some reason, the client takes too long in sending the keep alive
message and the firmware gets disconnected we wouldn't noticed it until
everything stops working. The most common reason for this to happen (at
least, in my experience) is dragging Squeak windows, which seems to stop
executing the keep alive process (regardless of its priority).