A few improvements and fixes to the communication protocol.
- Removed forking in UziProtocol>>#privateDisconnect. I added the forking
only in the places it was necessary.
- Added a few delays in UziProtocol>>#requestConnection. I'm not sure these
are needed, though. I was testing with a board that failed to connect from
time to time and it seemed to me that waiting a little before reading
helped. But I don't know. In any case it doesn't hurt to wait a little.
- Added a flag to UziSerialStream to log everything to the Transcript.