Simplified uzi.js implementation of the pin monitoring loop.
Now each website can have the set of pins it is interested in monitoring
and it will send them as a parameter on each request. No longer we will
handle activating/deactivating pin reports from the clients. That will be
done on the server side so that we can have multiple clients monitoring
different sets of pins without collisions.