Changed VM to store the halted script instead of the pc.
This is useful because the monitor needs to know the script in order to
send the data to the debugger. Also, I haven't confirmed this with a test
but I think storing the pc in which we last halted it's problematic in the
case where two or more taks are waiting to execute the same instruction.