Added a very ugly support to set breakpoints in the morphic UI.
I had to remove the stepping on UziDebuggerMorph and replaced it with
events instead because it was getting difficult to know when the update
was really needed. Before, the criteria was that the pc had changed but
now with breakpoints it could happen that the pc didn't change but an
update is needed anyways. It works much better now.