Changed UziDebugger to avoid sending so many messages.
It's a small improvement, though. Now instead of sending the breakpoints
over to the protocol immediately it will store the final state locally and
send them all at once when the user requires it (which is usually when they
click one of the debugger buttons). This way we avoid unnecessary
interactions.