Fixed bug that was considering script call instructions as statements.
They are not statements. All calls leave the return value on the stack.
If the value is discarded, then the compiler emits a pop after the call.
So that is the actual statement for the call.
I also changed the compiler so that the pop emitted will have the same
node as the expression whose value it is discarding. This way the
instruction is correctly highlighted.