Fixed bug in UziCallNode>>#hasSideEffects.
This method is called from the compiler in order to avoid emitting code to
perform short-circuit when there are no side effects. However, it was wrong
because it was ignoring the arguments and only taking into account the
primitive, which failed in cases like:

	!foo()

Here, the primitive is side-effect free but the argument is a script call
and we don't know if it's safe so we must consider it has side effects.