Changed UziProtocol and UziDebugger relationship.
Before, I created UziDebugger instances whenever I needed because each
debugger was associated with a running script. Now that all script halt
when a breakpoint is hit it doesn't make much sense. So I added a
"debugger" instance to UziProtocol and instead of updating the state via
events I can reference it directly.