Changed UziDebugger to rebuild its program on each access.
Before, I was initializing the "program" inst var at creation time because
I was assuming that the debugger instance was useful for a limited time.
Now that I keep an instance alive during the entire lifecycle of the
protocol I need to make it work even if the program changes. Of course, I
will need to detect when the vm is halted to avoid calling debugging
methods that are invalid for the vm state, but that's a problem for the
future.