Updated editor to show the current instruction when the debugger is active.
Also fixed a bug when setting the breakpoint that was sending the line #
instead of the pc. This should obviously be handled at the server side.
In fact, the web clients should not even talk about source code and line
numbers but about nodes in the AST. But for now, this works...