Changed the behavior of the KEEP_ALIVE to make it easier to debug the VM.
Debugging the VM was a pain in the ass because after every tick the VM was
disconnected. Now I added a keep alive counter that needs to go to 0
before the client is disconnected (and I decreased the interval,
accordingly). It should work exactly the same except when debugging, which
would give us several ticks before disconnecting (currently, 100 ticks).