Fixed a bug with the debugger.
When the VM was halted the tasks would unsync due to the next run being
dependant on the coroutine's last start. Now the last start is updated
when the VM tries to execute a coroutine but then it bails because it is
halted.
I'm not sure this implementation is actually correct but the observable
behavior makes much more sense now.