Changed colors and checkbox label.
The label was "Run automatically" and I changed it to "Interactive mode". I'm not sure this is a good description, though.
The colors in the "Pins" and "Globals" panel were changed to avoid using red to display change. Instead we'll use gray for the values that don't change.
Finally, the "Tasks" panel now uses an orange skull for the tasks with error.