Fixed potential bug in UziPinResource>>#detail:
When building the #history we were calling UziJSON>>#createObject: on all the pin values but we actually meant #createJSONObject:, which  properly encodes special floating point values (Infinity, -Infinity, and NaN).
It's just a typo, and the reason we didn't see the bug before is because pins never have those special values. Nonetheless, it was incorrect so I fix it.