Changed Sonar.uzi to store the distance in a global variable.
This will allow me to call the distance_xx() functions several times in a tick without worrying about interference. The downside is that I now need another task, and the setup code needs to start the "reading" task. Also, I made the task run 10/s which seems to be enough for most cases but I don't know...