Starting to change how the primitives are resolved.
Now the import resolver is also responsible for binding the calls to their respective primitives, when appropriate. This should give me some flexibility in the way I use the "prim" syntax to assign names to primitives. However, some tests are broken now and need to be fixed. And also, until I add the core primitives as implicit import I added a hack in UziProgramNode>>#primitiveForSelector: that simply ignores aliasing. It's ugly and should be removed ASAP.