Added UziParseNodeVisitorTest.
This class is responsible for detecting wether all subclasses of UziParseNodeVisitor implement all the visitor methods. This would be very useful whenever I add a new type of parse node or a new visitor.
Also, its test methods are autogenerated when I ran the tests so I don't need to update it manually, which is nice.