I decided to remove the "statements" input from the "declare_local_variable" block.
At first, I thought of making it work like the "for" block, that only allows to reference the local variable inside its body. But since the language syntax for variable declarations doesn't work like this I couldn't implement this the same way for the "declare_local_variable" without either changing the parser or adding a lot of checks in the blocks2ast code. So I decided to remove the "statements" input, which was the easier change. I'm not entirely happy about it but it's fine.