Fixed bug in the IDE startup.
Some blocks (like "start_task") require the uzi server connection established in order to access the current tasks. If the IDE is initialized before the connection was established and you happen to have one of those special blocks then the IDE would crash trying to build the program blocks. So now we wait for the uzi connection to be established before initializing the IDE