Big refactoring.
I moved the variables-in-scope and variable-name-and-scope from compiler.core to ast-utils. I realized there is a bug in the way the linker resolves variable scopes and the correct way to figure out if a given variable is local is to look at the path. Also, since variables-in-scope depends on the assign-internal-ids it seemed reasonable to move this function as well to ast-utils.
Finally, since the path is now required for these functions I changed the transform and transform-pred functions to pass the path as second parameter of all the clauses.