Changed parser output for floats to type Double.
This is not entirely correct because the uzi vm doesn't handle doubles but it makes testing easier because clojure's default type for decimals is double. And the responsibility for encoding the actual value falls in the encoder so it shouldn't be a problem if the parser and compiler treat decimal values as doubles.