Fixed subtle bug.
When compiling a pin literal node now we check against the specified board for its actual number. But if the pin is not valid then we get nil as a result and eventually the emitter tries to emit a constant with nil value, which throws a NullPointerException. Now I changed emitter/constant to use 0 as default value in case of nil. I'm not sure 0 is a good default, though, because we have a 0 pin in the Arduino. On the other hand, the UziFirmware ignores the pins 0 and 1 so there shouldn't be a problem (for now)