Changed uzi syntax toggle to be independent from i18n.
I removed the radio button from the "Internationalization" panel and instead added a checkbox in the "User interface" panel that controls a flag variable. The initBlock function now looks at this flag to decide wether to translate a block text or leave it as is. Also, the IDE now stores this flag as part of the settings.