Changed options dialog to show which panels are active.
I didn't implement yet the ability to enable/disable specific panels from the options dialog. To hide my incompetence I decided to simply disable the checkboxes and use them to display which panels are active at the moment. You can still reset the layout. It's not the best solution but for the moment it will do.