Small optimization of the compile stats process.
By changing the process in order to compile all the programs at the same time I can easily move the entire process to another thread. That saves a couple of seconds when running all the tests. I will have to eventually optimize the actual compilation process, but for now this will do.