Moved the port related code into its own file.
I also added a lot of logging code to try to understand why the programs gets stuck when opening/closing an unresponsive bluetooth port.

I seem to have been able to fix the open problem by making my own custom-open function (which is just a copy of serial.core/open that sets the timeout parameter to 0) and then calling that custom-open in another thread so that I could interrupt it if it gets stuck.

However, no matter what I did I can't seem to fix the close problem in a reliable way. I tried the same strategy of using a new thread and then interrupting it, I also tried just spawning a new thread and let it go. But even though I could regain control, in all my tests after closing the bluetooth port the application would not be able to connect to a valid port anymore, forcing me to close and restart the program.

IMPORTANT: I don't really know what's happening behind the scenes. I will probably have to debug the purejavacomm library. However, with this "fix" (which includes this commit and 128e39fa01f7e9ebaba5dbef5292f055571b499f ) the situation is worst than before. At least before the "connect" button would get permanently disabled, which would make it obvious that something wrong happened and force the user to restart the app. Now, the UI doesn't show any indication that the program is in an invalid state. So I think this branch should NOT be merged into master. At least until I find a more reliable solution.