Changed encoder related tests.
I don't want these tests to depend on the encoding details because it would make it difficult to change the encoder. Also, most of these tests were used in the clojure port to make sure the new code worked exactly as before. It should not be necessary now. I kept the tests anyway because they exercise the compiler and encoder. And also because they supply programs to the CompileStats.csv file