Reduced monitor and server delays
This is just a test to see if I can reduce the latency.
The VM report interval seems to work fine but in detriment of overall performance. Probably a dynamic interval would be better. Although I don't know yet how should I do it.
Reducing the middleware timeout doesn't seem to be a problem at all.