Added a small configuration module.
For now, it only allows to configure the device report-interval but eventually will allow to control other parameters.
The implementation is very simple: it reads a file config.edn and it caches its value until the file is modified. If the file is not found then the read-config function returns nil. It's the caller responsibility to handle this with default values.
For development the file is placed in the project root folder. For builds, this file needs to be placed in the same directory as the start scripts.