Added the Uzi.setProfile(..) function in the client.
This way I can toggle the profiling from the browser and I don't need the REPL, which takes too long on the orange pi.
I also removed a couple of logger calls and replaced them with pseudo vars. So now the ticks per second and the vm's report interval are visible in the inspector.