Changed monitor to have two separate variables for the report interval.
One represents the lower bound set by the middleware, and the other is the actual report interval set by the VM.
This way, if the middleware is slow it can specify a report interval, but if this value is still too much for the VM and performance degrades it can still increase it to reach its goal.
This change allowed me to remove the fixedReportInterval flag, and I think the code is a little better now.
Finally, I should mention that now the report interval uses only 7 bits (before it was 13 bits) so the largest interval possible is 127.