Added a temporary hack to avoid refactoring the compiler.
The problem I have is that the linker needs access to the file system in order to read the imported libraries. However, when running in the browser I don't have direct access to the file system. Instead I need to go through an http request. That should be straightforward but the async nature of javascript makes it a little annoying because I need to change the entire compiler code to make it asynchronous as well, which I don't really feel like doing right now. So as a temporary solution I wrote a macro that reads all the libraries at compile time and leaves them stored in the code. Since clojurescript macros run in the context of the JVM, I can access the filesystem without problems. And this way the compiler can keep its sync implementation. It's an ugly hack and I will fix it soon, but for now, it seems to work.