Fixed small problem with the protocol tests.
These tests are currently testing the controller, which I'm in the process of refactoring to extract the protocol code out into its actual namespace. So the problem was that some tests rely on the controller state being set to a particular program. When the refactor is complete this will no longer be necessary and I'll simplify this code.