Improved thread-safety of controller's connect! and disconnect! functions.
Since both functions take some time to finish, it was possible to trigger multiple connection/disconnection attempts if they were done simultaneously.
To prevent potential inconsistencies I added a :pending state to the connection. The :pending state means either a connection or disconnection is in process, in which case all other attempts will fail until the current process finishes (successfully or not).
The code is kind of messy right now because I've added a few log calls to help me test. I'll clean it later.