Splitted cljs test build in two: compiler-test and controller-test.
Before, all tests ran on the browser. Now the compiler-test run in the browser and controller-test run in node.
I also changed the compiler tests to avoid running the vm-test-files-gen namespace. Since I can now run tests in node and I have the fs abstraction, I will be able to actually generate the files instead of the current hack.