Optimization. Moved AST checking step to last (after dead code removal).
This way the AST only checks actual code used in the program and not the expanded AST with all the imports resolved.
Also, for this to work I had to change the dead-code-remover which was removing the imported program from the import node so the check was then failing because it was impossible to find the declared variables (because their creation scope was lost).
I think this change is pretty safe but I'm thinking maybe the linker could attach metadata to the imports instead of an actual key.