Implemented compile in the simulator code.
This means we can simplify the uzi.js code and also remove the compiler build.
We'll have to change how we detect if we're running locally or not. Before I was using the presence of the middleware namespace but if I want the simulator to be present always I need to change how this works.
Maybe always compile locally? That's probably the best solution but we need to make sure the local compilation is as fast as the server.