Fixed bug in the simulator step.
I was only checking the serial after all the loops were done. However, if a lot of time had passed (maybe because the javascript window lost focus and didn't execute the timers on time) then we could end up executing a lot of loops in one timer tick. This could result in the serial buffer overflowing, which would trigger a bunch of weird errors with the communication. Now we read the serial every 10 loops, which should be safe.