Experimenting with how the core update-loop works.
I added this function that will use the controller's update mult but then I realized that I was getting updates much faster than I wanted, so I changed it to send updates to the clients with a 50 ms interval. Even though the code is much more complicated now I think it should be more efficient because we're not building the server state map on every iteration just to be able to detect the changes. However, we're still spinning every 50 ms which is not ideal. I would prefer to have the update loop wake up whenever there are changes and then go back to sleep but still limit the update rate to 50 ms.