Changed ast.utils/transform to avoid adding the current node to the path.
The problem was that nodes were being added to the path before being transformed, now we wait until after evaluating the clauses to add the node to the path. This way children will get the transformed parent in their respective path. If I'm not mistaken this should allow us to safely compare nodes by identity and thus not rely on the annoying :internal-id hack.