Fixed 2 compiler bugs and added tests.
The first bug was pretty mild. The compiler was throwing an "Unknown node" exception when calling a func/proc with less named arguments than required.
The second bug was very bad. When compiling a call with less positional arguments than required, the compiler was not throwing any exception (which is good) but it was not emitting the instructions to push default values for the missing arguments. That meant that when executing the program it could mess up the stack. Very bad stuff.