Fixed bug when disconnecting the simulator.
It was not resetting the controller state to its initial state.
I change this by mistake when refactoring the :program stuff because before we were preserving the :current program and since :current program was always nil I removed the entire line without realising this was setting the state to initial-state as well.