Implemented a very basic "read-only" mechanism for the block editor.
I simply store the XML program at the time of setting the readOnly flag. And then whenever a user interaction provokes a change in the workspace I just revert it back to the stored program. It seems to work fine but I'll have to test it more to be sure. Interacting with blockly can give some unpleasent surprises...