Fixed bug in the AST -> block generation of script calls.
The problem was that the language allows for both positional as well as named arguments. However, the block generation for calls was assuming always positional arguments. For most calls this was not an issue, but for some it could lead to either wrong code being generated or unnecessary HERE_BE_DRAGONS.