Refactoring compiler.
I simplified the ast-utils/transform functions by removing the clause evaluation and instead just passing a transforming function that should perform all the necessary checks inside.
I also renamed transform to transform-with-path and added a separate version that doesn't keep track of the path of parent nodes.
Apart from being simpler this code should also be slightly faster. Some basic profiling shows a speedup beetween 10% and 15% on the whole compilation. It's not revolutionary but it's something.