Fixed inspector bug when disconnecting.
Since we reset the state after a disconnect, we lose the reporting pins/globals. However, we were not triggering the update on the pins and globals for the IDE so the inspector was still under the wrong impression that some pins/globals were being reported. This meant that the inspector will show stale data and it wouldn't be obvious to the user that the data is not being updated.