Fixed bug with encoding of branch instructions.
I forgot I used 1 bit of the instructions data to store the breakpoint flag, which cuts in half the length of the jumps we can perform. I found this now because we started using the regular "task" with a "forever" block inside instead of the "timer" block, which made the problem more obvious as we now have a jmp instruction at the end of the task. To fix it I added a bunch of primitives that essentially perform the same jumps but taking the offset from the stack instead of the instruction's data. This means we can perform much larger jumps at the cost of a bigger program. To avoid having to penalize all programs with this change I tried to detect when the jump offset is not encodable in 7 bits and compile the primitive version. The code is just a big mess that I will fix as soon as I can, but it seems to work ok. I also added a lot of VM tests to verify.