Fixed tests in vm_test_files_gen.
I mostly had to replace a lot of prim-call with script-call. Then, some of these script-call were being treated as statements so I had to add a "pop" afterwards (this isn't necessary for prim-calls because the compiler knows the stack-transition for each prim). I had to fix a few issues with the logical and/or instructions because the compiler tries to optimize this calls if it know the right doesn't have side effects (which all script calls are treated as having side effects so I had to fix all these cases as well). Finally, since most of these programs were emitted manually I had to add a method that compiles the core.uzi and injects the required scripts into the program, otherwise the script-call won't find the target script.