Updated project.clj.
Moved the ultra plugin to the :test profile and added humane-test-output to the :dev profile. The problem was that when running the tests without the terminal the ultra plugin was adding extra characters for the colors. So I took out the ultra plugin from :dev but added the humane-test-output because of the better diffing.