Fixed VM bug when returning from a script call.
The problem was that now that the compiler doesn't emit the "pop" instructions when calling scripts as statements we can have a script whose last instruction is a script call. This shouldn't be an issue except for the fact that when executing the SCRIPT_CALL instruction we're pushing the current pc to the stack so that the VM knows where to return. This pc points to the next instruction to be executed, which may be outside the limits of the calling script, as that's perfectly legal: the VM will then handle the situation as if a return was called (depending on the type of script). The problem is that the VM needs to know to which script to return in order to decide how to proceed, but the way we were finding the script was by using the return pc. If the call was done as the last instruction of the script then the pc will likely point to the next script in the program or to a non existing script. In both cases it's a disaster for the VM because the actual script is impossible to find and thus we can't properly execute the return.
A potential solution to this problem would be to detect these cases in the compiler and emit a NOP operation after the call to simply avoid this situation altogether. I don't like this because it needlessly increases the size of the programs. I also don't like having to introduce a new instruction that does nothing just for this. Maybe I would be able to use the existing PRIM_RET instead of a new NOP instruction. However, I still don't like to increase the size of user programs.
So the solution I decided to implement was to change the meaning of the pc we push to the stack when calling a script. Instead of the next instruction now we push the pc of the currently executing instruction, the SCRIPT_CALL, which we know for sure is in the correct script and thus will allow us to set the currentScript property correctly on return. To do this I simply decrement the pc before pushing it into the stack. And I changed the VM::unwindStackAndReturn(void) function to set the currentScript using the pc from the stack and then increment it before returning.
There is 1 exception, though, if the framePointer is -1 (which means that we're starting the execution of the task) we don't decrement the pc. It's a special case and I don't really like it but it kind of makes sense because if we decrement it anyways we would have the same problem we're trying to solve: the pc would be off by 1 and it will point outside of the script but in the opposite direction.
So anyways, I decided to live with this hack for the time being. It seems to work fine. I don't exactly remember why I need to initialize the stack with -1 for the framePointer when we're starting the execution but I don't have enough brainpower to find out right now, and also I don't want to risk subtle bugs so for now we'll keep that as is.