Changed program.utils/statement? to handle script calls correctly.
Before, all script calls were considered expressions because the VM would always leave a value on the stack. However, now we need to differentiate function calls from procedure/task calls to know if the call is a statement or not.
This is important for the debugger to properly analyze the program and group the instructions correctly.