Fixed debugger bug when stepping over a return.
I think the problem is well explained in the code comment but just as a brief recap: we can't use the pc in the stack frame because it now points to the call and we need the next instruction, and we can't increment it willy-nilly because we might end up outside the calling script. So what we do is simply walk the stack and find all the returns that, after being incremented, still lie within the calling script. That's it. I'm not sure it's the correct solution in all cases but it seems to work fine.