Fixed bug in debugger/step-into-call.
This bug happened when the instruction group contained both a call and then a branch, and the current pc was pointing to the branch. In that case, since the call was executed already it means we came back from the script and we need to put breakpoints on all potential branches. However, the fallback code was step-over-regular, which assumes no branches. Instead I just call step-over, which should handle all cases correctly.