Fixed bug in the cljs build hook.
I wanted to only copy the main.js and manifest.edn file but file-seq is recursive and traverses the entire tree. If there were other files in there, like the ones generated by the compiler in debug mode, then they would get copied as well. The simple fix is to just replace file-seq with listFiles.