Fixed bug that was assuming some libs variables were always init as pins.
Usually, libraries such as DCMotor or Sonar would have their variables initialized as pins because the UI forces their initial value to be a pin. However, if the program was originally written by code then this no longer is a valid assumption. In those cases the blocks2ast generation was failing because the code always tries to generate a pin node when the value does not conform to the pin syntax. Now we are more flexible: we try to generate a pin node but if the syntax is not valid we generate a number node.