Fixed compilation of variable declarations.
Since variables have a default value, there was an optimization that prevented the compiler from emitting the assignment for a local variable if the variable was already set to a compile-time constant. This was incorrect because if the variable declaration was inside a loop then the declaration should reset the value of the variable on each iteration. However, this was not happening so now I fixed it by checking that the ctx path doesn't contain a loop node: if it does we emit the assignment regardless of the value being a compile-time constant or not.