Improved code for keeping the top blocks in their place when changing the code.
The main idea is that instead of saving the absolute positions of each block we now store the relative position to the previous block's bottom-left. This not only allows to simplify the code substantially but also fixes a problem when updating the code inside a block: in the old version the new blocks could overlap with the existing blocks below, but now the existing blocks will move down to make room.